elbepack: tests: introduce build_driver abstraction for xml tests
Split the test_xml to only contain code that is not directly related to
the initvm. This will enable us in the next step to add further tests
that will use the container-based approach.
