+++ elbepack/tests/test_xml.py
@@ -2,107 +2,41 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # SPDX-FileCopyrightText: 2020 Linutronix GmbH
 
-import contextlib
-import functools
-import io
+
+# Only contains templates used by other test files
+__test__ = False
+
 import pathlib
-import subprocess
 
 import pytest
 
 from elbepack.main import run_elbe_subcommand
-from elbepack.tests import parametrize_xml_test_files, xml_test_files
+from elbepack.tests import xml_test_files
 
 
-here = pathlib.Path(__file__).parent
+CHECK_BUILD_VARIANTS = ('schema', 'cdrom', 'img', 'sdk')
 
-
-@pytest.fixture(scope='module')
-def initvm(tmp_path_factory, request):
-    initvm_dir = tmp_path_factory.mktemp('initvm-') / 'initvm'
-    use_initvm = request.config.getoption('--elbe-use-initvm')
-
-    if use_initvm in {'libvirt', 'existing'}:
-        qemu_arg = []
-    elif use_initvm == 'qemu':
-        qemu_arg = ['--qemu']
-    else:
-        raise ValueError(use_initvm)
-
-    def initvm_func(subcmd, *args):
-        run_elbe_subcommand(['initvm', subcmd, '--directory', initvm_dir, *qemu_arg, *args])
-
-    def destroy_initvm():
-        with contextlib.suppress(Exception):
-            initvm_func('stop')
-        with contextlib.suppress(Exception):
-            initvm_func('destroy')
-
-    if use_initvm == 'existing':
-        yield initvm_func
-        return
-
-    try:
-        initvm_func('create', '--fail-on-warning')
-    except Exception as e:
-        # If the fixture setup fails, pytest will try to create the fixture for
-        # each test. This is very slow and unlikely to work, so remember the failure.
-        def error_func(*args, _initvm_exception, **kwargs):
-            raise RuntimeError('initvm setup failed') from _initvm_exception
-        destroy_initvm()
-        yield functools.partial(error_func, _initvm_exception=e)
-    else:
-        try:
-            yield initvm_func
-        finally:
-            destroy_initvm()
-
-
-def _delete_project(uuid):
-    with contextlib.suppress(Exception):
-        run_elbe_subcommand(['control', 'del_project', uuid])
+_EXTENDED_XML = (
+    pathlib.Path('tests') / 'base-extended' / 'simple-validation' / 'image-extended.xml'
+)
 
 
 @pytest.fixture(scope='module', params=xml_test_files('simple'), ids=lambda f: f.name)
-def simple_build(request, initvm, tmp_path_factory):
-    build_dir = tmp_path_factory.mktemp('build_dir')
-    prj = build_dir / 'uuid.prj'
-
-    initvm(
-        'submit', request.param,
-        '--output', build_dir,
-        '--keep-files', '--build-sdk',
-        '--writeproject', prj,
-    )
-
-    uuid = prj.read_text()
-
-    with contextlib.redirect_stdout(io.StringIO()) as stdout:
-        run_elbe_subcommand(['control', 'list_projects'])
-
-    if uuid not in stdout.getvalue():
-        raise RuntimeError('Project was not created')
-
-    yield build_dir
-
-    _delete_project(uuid)
+def simple_build(request, tmp_path_factory, build_driver):
+    workdir = tmp_path_factory.mktemp('build_dir')
+    return build_driver.submit(request, request.param, workdir, build_sdk=True)
 
 
 @pytest.mark.slow
-@pytest.mark.parametrize('check_build', ('schema', 'cdrom', 'img', 'sdk'))
+@pytest.mark.parametrize('check_build', CHECK_BUILD_VARIANTS)
 def test_simple_build(simple_build, check_build):
     run_elbe_subcommand(['check-build', check_build, simple_build])
 
 
 @pytest.mark.slow
-def test_rebuild(initvm, simple_build, tmp_path_factory):
+def test_rebuild(build_driver, simple_build, tmp_path_factory):
     build_dir = tmp_path_factory.mktemp('build_dir')
-
-    initvm(
-        'submit', '--skip-build-source',
-        '--output', build_dir,
-        simple_build / 'bin-cdrom.iso',
-    )
+    build_driver.rebuild(simple_build / 'bin-cdrom.iso', build_dir)
 
 
 @pytest.mark.slow
@@ -110,65 +44,17 @@ def test_check_updates(simple_build):
     run_elbe_subcommand(['check_updates', simple_build / 'source.xml'])
 
 
-def _prjrepo_list_packages(uuid):
-    with contextlib.redirect_stdout(io.StringIO()) as stdout:
-        run_elbe_subcommand(['prjrepo', 'list_packages', uuid])
-
-        return stdout.getvalue()
-
-
 @pytest.mark.slow
-@parametrize_xml_test_files('xml', 'pbuilder')
-def test_pbuilder_build(initvm, xml, tmp_path, request):
-    build_dir = tmp_path
-    prj = build_dir / 'uuid.prj'
-
-    run_elbe_subcommand(['pbuilder', 'create', '--xmlfile', xml, '--writeproject', prj])
-
-    uuid = prj.read_text()
-    request.addfinalizer(lambda: _delete_project(uuid))
-
-    # Not necessary, to test the command.
-    run_elbe_subcommand(['pbuilder', 'update', '--project', uuid])
-    run_elbe_subcommand(['control', 'wait_busy', uuid])
-
-    assert _prjrepo_list_packages(uuid) == ''
-
-    for package in ['libgpio', 'gpiotest']:
-        subprocess.run(['git', 'clone', f'https://github.com/Linutronix/{package}.git'],
-                       check=True, cwd=build_dir)
-        run_elbe_subcommand(['pbuilder', 'build', '--project', uuid,
-                             '--source', build_dir.joinpath(package),
-                             '--output', build_dir.joinpath('out')])
-
-    assert _prjrepo_list_packages(uuid) == (
-        'gpiotest_1.0_amd64.deb\n'
-        'libgpio-dev_3.0.1_amd64.deb\n'
-        'libgpio3-dbgsym_3.0.1_amd64.deb\n'
-        'libgpio3_3.0.1_amd64.deb\n'
-    )
-
-    run_elbe_subcommand(['prjrepo', 'upload_pkg', uuid, here / 'equivs-dummy_1.0_all.deb'])
-
-    assert _prjrepo_list_packages(uuid) == (
-        'equivs-dummy_1.0_all.deb\n'
-        'gpiotest_1.0_amd64.deb\n'
-        'libgpio-dev_3.0.1_amd64.deb\n'
-        'libgpio3-dbgsym_3.0.1_amd64.deb\n'
-        'libgpio3_3.0.1_amd64.deb\n'
-    )
-
-
-@pytest.mark.slow
-def test_base_extended_build(simple_build, initvm, tmp_path):
-    tests_dir = pathlib.Path('tests') / 'base-extended' / 'simple-validation'
-    extended_xml_path = tests_dir / 'image-extended.xml'
+def test_base_extended_build(request, build_driver, simple_build, tmp_path):
     base_build_image = simple_build / 'base-rootfs.tgz'
-    extended_build = tmp_path / 'extended-build'
 
     if not base_build_image.exists():
         pytest.skip('No base image tarball was produced')
 
-    initvm('submit', '--output', extended_build, '--skip-build-bin', '--skip-build-sources',
-           '--base-image', base_build_image, extended_xml_path)
-    run_elbe_subcommand(['check-build', 'img', extended_build])
+    extended_build = tmp_path / 'extended-build'
+    extended_build.mkdir()
+    build_dir = build_driver.submit(
+        request, _EXTENDED_XML, extended_build,
+        skip_cdrom=True, base_image=base_build_image,
+    )
+    run_elbe_subcommand(['check-build', 'img', build_dir])
